+++ apps/ui-dev/component/patterns/panel/ToolView.jsx
@@ -1,0 +1,29 @@
+import MoreHorizFilled from "@elf-framework/icon/MoreHorizFilled";
+import { Button, Panel, VBox } from "@elf-framework/ui";
+
+export function ToolView() {
+  return (
+    <VBox
+      style={{
+        gap: 30,
+        backgroundColor: "var(--color-gray-0)",
+        padding: "30px 200px",
+        fontSize: 13,
+        justifyContent: "center",
+        alignItems: "center",
+      }}
+    >
+      <Panel
+        title={"fdsf"}
+        tools={[
+          <Button iconOnly>
+            <MoreHorizFilled />
+          </Button>,
+        ]}
+      >
+        Learn how to work with green screen video with these tips for video
+        makers.
+      </Panel>
+    </VBox>
+  );
+}