+++ apps/ui-dev/layouts/Layout.jsx
@@ -39,19 +39,18 @@ export function Layout(props) {
         </Flex>
       </div>
       <View class="layout-menu">
-        {version ? (
-          <Flex style={{ paddingBottom: 30 }}>package: &nbsp; {version}</Flex>
-        ) : undefined}
+        <PageMenu
+          menu={menu}
+          header={version ? <Flex>package: &nbsp; {version}</Flex> : undefined}
+          footer={
+            <ActionGroup justified>
+              <Button as="link" href="https://github.com/elf-framework/editor">
+                Github
+              </Button>
+            </ActionGroup>
+          }
+        />
 
-        <PageMenu menu={menu} />
-        <Blank style={{ height: 100 }} />
-        <div>
-          <ActionGroup justified>
-            <Button as="link" href="https://github.com/elf-framework/editor">
-              Github
-            </Button>
-          </ActionGroup>
-        </div>
         <Blank />
       </View>
       <View class="layout-content">