+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/ui
 
+## 0.0.101
+
+### Patch Changes
+
+- update panel component
+
+- Updated dependencies []:
+  - @elf-framework/design-system@0.0.101
+
 ## 0.0.100
 
 ### Patch Changes